Save progress LDS resource not found doesn't work
