+++ internal/internal.go
@@ -188,7 +188,7 @@ var (
 	// error for a given resource type and name. This is usually triggered when
 	// the associated watch timer fires. For testing purposes, having this
 	// function makes events more predictable than relying on timer events.
-	TriggerXDSResourceNameNotFoundForTesting any // func(any, string, string) error
+	TriggerXDSResourceNameNotFoundForTesting any // func(func(xdsresource.Type, string), string, string) error
 	// TriggerXDSResourceNameNotFoundForTesting triggers the resource-not-found
 	// error for a given resource type and name through an xDS Server. This is
 	// usually triggered when the associated watch timer fires. For testing
