+++ xds/internal/xdsclient/singleton.go
@@ -20,45 +20,20 @@ package xdsclient
 
 import (
 	"fmt"
-	"google.golang.org/grpc/xds/internal/xdsclient/xdsresource"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	"google.golang.org/grpc/internal"
 	"google.golang.org/grpc/internal/envconfig"
 	"google.golang.org/grpc/internal/grpcsync"
 	"google.golang.org/grpc/xds/internal/xdsclient/bootstrap"
 )
 
-func init() {
-	internal.TriggerXDSResourceNameNotFoundClient = triggerXDSResourceNameNotFoundClient
-}
 
 const (
 	defaultWatchExpiryTimeout         = 15 * time.Second
 	defaultIdleAuthorityDeleteTimeout = 5 * time.Minute
 )
-
-func triggerXDSResourceNameNotFoundClient(resourceType, resourceName string) error {
-	singletonMu.Lock()
-	defer singletonMu.Unlock()
-	/*var typ xdsresource.Type
-	switch resourceType {
-	case xdsresource.ListenerResourceTypeName:
-		typ = listenerType
-	case xdsresource.RouteConfigTypeName:
-		typ = routeConfigType
-	case ClusterResourceTypeName:
-		typ = clusterType
-	case EndpointsResourceTypeName:
-		typ = endpointsType
-	default:
-		return fmt.Errorf("unknown type name %q", typeName)
-	}*/
-	return internal.TriggerXDSResourceNameNotFoundForTesting.(func(xdsresource.Producer, string, string) error)(singletonClient, resourceType, resourceName)
-	// singletonClient.triggerResourceNotFoundForTesting()
-}
 /*
 func singletonClientRef() any/*xdsresource.Producer { // can import xdsresource since doesn't create cycle
 	singletonMu.Lock()
@@ -123,6 +98,7 @@ func newRefCountedWithConfig(fallbackConfig *bootstrap.Config) (XDSClient, func(
 	if err != nil {
 		return nil, nil, err
 	}
+	print("setting singleton client") // does it not get here in tests?
 	singletonClient = &clientRefCounted{clientImpl: c, refCount: 1}
 	singletonClientImplCreateHook()
